MODIFIED: refactored align to use sapi
